+++ tests/data/EmptyFactory.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Test\Ebln\PHPStan\EnforceFactory\data;
+
+use Test\Ebln\PHPStan\EnforceFactory\data\code\EmptyProduct;
+
+class EmptyFactory
+{
+    public function class(): EmptyProduct
+    {
+        return new EmptyProduct();
+    }
+}