+++ CHANGELOG.md
@@ -5,8 +5,14 @@ Intended to follow [«Keep a Changelog»](https://keepachangelog.com/en/)
 
 ----
 
+## Upcoming
+
+- Support for attributes
+
+----
+
 ## [0.0.1]
 
 ### Added
 
-- Rule!
+- Rule to enforce Factory pattern via interface